teljesen kész

+++ gubge_beadando/src/points.hpp
@@ -8,9 +8,9 @@
 #ifndef POINTS_HPP_
 #define POINTS_HPP_
 struct points{
-	int x,y;
+	double x,y;
 	points(): x(0),y(0){};
-points (int a,int b): x(a),y(b){};
+points (double a,double b): x(a),y(b){};
 };
